requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
pg8000==1.30.5
openai==1.3.0
httpx==0.24.1
python-dotenv==1.0.0
Flask-Login==0.6.3
Flask-WTF==1.1.1
WTForms==3.0.1
Werkzeug==2.3.7
requests==2.31.0
